Added a Search Controller, for the Search view. Also created a UML Class diagram, that reflects the class structure. Made modifications to fxml files to meet requirements.

+++ src/application/PriceWatcherModel.java
@@ -6,14 +6,25 @@ public class PriceWatcherModel extends ScrapePrice{
 
     static Item [] items = new Item[10];
 
+    /**
+     * Uses random values to return a price change.
+     * @param price
+     * @return price change
+     */
     public static float calculateChange(double price){
         double randPrice = Math.random();
         float percentage = (float) ((price*10)/ randPrice);
         //float percentage = (float) ((price)/ price) - 1;
         return percentage;
     }
+
+    /**
+     * Calls web scraper to return price
+     * @param url
+     * @return price
+     */
     public double getPrice(String url){
-        return ScrapePrice.priceScraper(url);
+        return (double) ScrapePrice.priceScraper(url);
     }
 
     public static Item addItem(String name, double price, String url){
